FS class

FS class

Arguments

server_address

placehoder

api_address

placehoder

vsfs_jar

placehoder

cache_dir

placehoder

cache_size

placeholder

project_id

placeholder

Methods

file(id = NULL)

given project id, show all files in it

mount( mount_point = NULL, project_id = NULL, ignore.stdout = TRUE, sudo = TRUE, ... )

mount a specific project if project_id is provided, otherwise mount all projects

path(id = NULL)

List path for all mounted projects, for easy copy/paste of file path. If project id is provoded, show project path and files path.

unmount(mount_cmd = NULL, project_id = NULL, ...)

unmount a project if project_id is provided, otherwise unmount all