SchemaList

SchemaList

A schema defines a parameter type.

SchemaDef Class

SchemaDefList

InputSchema Class

OutputSchema Class

SchemaList(...)

SchemaDefList(...)

Arguments

…

element or list of the element.

Value

a Schema object or sbuclass object.

Fields

type

(ANY) The data type of this parameter.

fields

[SchemaList] When type is record, defines the fields of the record.

symbols

[character] When type is enum, defines the set of valid symbols.

items

[ANY] When type is array, defines the type of the array elements.

values

[ANY] When type is map, defines the value type for the key/value pairs.

inputBinding

[Binding] Describes how to handle a value in the input object convert it into a concrete form for execution, such as command line parameters.

Examples

Schema(fields = SchemaList(SchemaDef(name = "schema")))