Reports the completion of the part upload

The ETag is provided for the correctness check upon completion of the whole upload. Value for the ETag is provided by AWS S3 service when uploading the file in the ETag header.

upload_complete_part(token = NULL, upload_id = NULL, part_number = NULL,
  e_tag = NULL, ...)

Arguments

token
auth token

upload_id
ID of the upload

part_number
ID of the part you wish to report as completed

e_tag
Value of the ETag header returned by AWS S3 when uploading part of the file.

...
parameters passed to api function

Value

parsed list of the returned json

Examples

# NOT RUN {
token = "your_token"
# }# NOT RUN {
req = upload_complete_part(token,
      upload_id = "8D7sQJxQk14ubsEnKaoeQZlRvV6ouQtMzBWaQNJdxPDLypUC3WogwtJdncevHxnT",
      part_number = "1",
      e_tag = "d41d8cd98f00b204e9800998ecf8427e")
# }