Class Folders

Class Folders

Value

Folders object

Details

Folders class is usally returned by the API call which returns Folders. It allows you to create and organize folders on platform. Folders are created inside projects, and can be used to contain files within the project. Users do not usually need to construct Folders manually, they are generated from a API call most of the time.

Fields

id

character string used as folder ID

name

character string used as folder name

project

project name the folder belongs to

parent

parent folder ID, usually returned by the project details (root_folder) or listed folder contents

type

file type: FILE or FOLDER

created_on

date the folder was created

modified_on

date the folder was modified

Methods

copy_file_to_folder(...)

copy a file between folders

create_folder(name = NULL, ...)

create a new folder

delete(...)

delete the folder

list_contents(...)

list folder contents

move_file_to_folder(...)

move a file between folders

Examples

# NOT RUN {
Folders(id = "test_id", name = "test.bam")
# }