List projects using Seven Bridges command line uploader

List projects available as upload targets using Seven Bridges command line uploader.

cli_list_projects(token = NULL, uploader = NULL, proxy = NULL)

Arguments

token
Authentication token.

uploader
The directory where Seven Bridges command line uploader is located (the directory that contains the bin/ directory).

proxy
A proxy server through which the uploader should connect. For details the proxy parameter format, see the part on parameter --proxy in the reference below.

Value

Character vector of the available project names.

References

https://docs.sevenbridges.com/docs/upload-via-the-command-line

See also

Examples

# NOT RUN {
token <- "your_token"
# }# NOT RUN {
cli_list_projects(
  token = token,
  uploader = "~/Downloads/sbg-uploader/"
)
# }