Upload files using Seven Bridges command line uploader

Upload files using Seven Bridges command line uploader.

cli_upload(token = NULL, uploader = NULL, file = NULL, project = NULL,
  proxy = NULL, tag = NULL, manifest_file = NULL,
  manifest_metadata = c("all", "none", "partial"), metadata_fields = NULL,
  dry_run = FALSE, dry_run_fields = NULL)

misc_upload_cli()

Arguments

token

Authentication token.

uploader

The directory where the command line uploader is located (the directory that contains the bin/ directory).

file

The location of the (single) file to upload. To upload multiple files, please use manifest_file to specify.

project

The project identifier (e.g. username/project-name) to upload files to. This option is mandatory. To upload files to a project, you must be a member of that project and must have the write permission granted by the project administrator.

proxy

A proxy server through which the uploader should connect. For details the proxy parameter format, see the part on parameter --proxy in the reference below.

tag

Tags for your the files (optional). Use a vector of character strings, for instance, c("tag one", "the second tag").

manifest_file

Location of the manifest file (for uploading multiple files with metadata). See the reference URL below for the format of a manifest file.

manifest_metadata

Should we use all, none, or only a part of the the metadata fields included in the manifest file? Default is "all".

metadata_fields

Character vector, the metadata fields to use in the manifest file. This should be specified if and only if manifest_metadata = "partial".

dry_run

Should we just output the data and check the settings without uploading anything? Default is FALSE.

dry_run_fields

Character vector, specific metadata fields to output information about when dry_run = TRUE.

Value

The uploaded file's ID number.

Note

To use the command line uploader, Java 1.7 or newer should be installed. See the reference link below for details.

References

Seven Bridges Command Line Uploader: http://docs.sevenbridges.com/docs/upload-via-the-command-line Manifest file format: http://docs.sevenbridges.com/docs/format-of-a-manifest-file

See also

See get_uploader for downloading the command line uploader for Seven Bridges platforms. See cli_list_projects and cli_list_tags for listing available projects or tags with the command line uploader.

Examples

token = "your_token"
cli_upload(
  token = token,
  uploader = "~/Downloads/cgc-uploader/",
  file = "~/example.fastq", project = "username/project-name")